Move dispatch_setup.c into dispatch.c
Also refactor some of the functions in dispatch.{c,h}
