+++ package.xml
@@ -59,7 +59,6 @@
                     <file name="vendor_stdatomic.h" role="src" />
                     <file name="dispatch.c" role="src" />
                     <file name="dispatch.h" role="src" />
-                    <file name="dispatch_setup.c" role="src" />
                     <file name="dogstatsd_client.c" role="src" />
                     <file name="dogstatsd_client.h" role="src" />
                     <file name="engine_hooks.c" role="src" />
