+++ src/ext/php5_4/dispatch.c
@@ -19,7 +19,7 @@ ZEND_EXTERN_MODULE_GLOBALS(ddtrace)
     } while (0)
 #endif
 
-zend_function *ddtrace_function_get(const HashTable *table, zval *name) {
+zend_function *ddtrace_ftable_get(const HashTable *table, zval *name) {
     char *key = zend_str_tolower_dup(Z_STRVAL_P(name), Z_STRLEN_P(name));
 
     zend_function *fptr = NULL;
